requirements.txt
# UI框架
PySide6>=6.5.0
qt-material>=2.14

# 动画效果
qtawesome>=1.3.0  # 图标支持

# 配置管理
pyyaml>=6.0.1

# 开发工具(可选)
black>=23.0.0
pylint>=2.17.0

# 其他依赖
opencv-python>=4.5.0 